This is getting seriously fun

+++ modules/backends/HTML5.js
@@ -1,15 +1,18 @@
 'use strict';
 
+let nextId = 0;
+
 export default class HTML5Backend {
   constructor(actions) {
     this.actions = actions;
+    this.mime = `application/dnd-${nextId++}`;
   }
 
   getSourceProps(sourceHandle) {
     return {
       draggable: true,
       onDragStart: (e) => this.handleDragStart(e, sourceHandle),
-      onDragEnd: (e) => this.handleDragEnd(e)
+      onDragEnd: (e) => this.handleDragEnd(e, sourceHandle)
     };
   }
 
@@ -23,26 +26,43 @@ export default class HTML5Backend {
   }
 
   handleDragOver(e) {
-    if (e.preventDefault) {
-      e.preventDefault(); // Necessary. Allows us to drop.
+    if (e.dataTransfer.types.indexOf(this.mime) === -1) {
+      return;
     }
 
-    e.dataTransfer.dropEffect = 'move';  // See the section on the
+    if (e.preventDefault) {
+      e.preventDefault();
+    }
+
+    e.dataTransfer.dropEffect = 'move';
   }
 
   handleDragEnter(e, targetHandle) {
+    if (e.dataTransfer.types.indexOf(this.mime) === -1) {
+      return;
+    }
+
     this.actions.enter(targetHandle);
   }
 
   handleDragLeave(e, targetHandle) {
+    if (e.dataTransfer.types.indexOf(this.mime) === -1) {
+      return;
+    }
+
     this.actions.leave(targetHandle);
   }
 
-  handleDrop() {
+  handleDrop(e) {
+    if (e.dataTransfer.types.indexOf(this.mime) === -1) {
+      return;
+    }
+
     this.actions.drop();
   }
 
   handleDragStart(e, sourceHandle) {
+    e.dataTransfer.setData(this.mime, {});
     this.actions.beginDrag(sourceHandle);
   }
